refactor: improve codebase architecture with ports/adapters and module isolation
- Relocate executeAuditAndRevalidate from moderation to audit module
- Isolate thread creation from infrastructure via ThreadSideEffectsPort
- Decompose 570-line ai.worker.ts monolith into 8 focused handler files
- Expand MessageSideEffectsPort to route all side effects through adapters
- Remove direct cross-module imports (notifications, activity, queue)
- All 347 tests pass, typecheck clean
