+++ lib/queue/workers/index.ts
@@ -1,0 +1,8 @@
+export { handleThreadSummaryJob } from './thread-summary.worker';
+export { handleThreadDnaJob } from './thread-dna.worker';
+export { handleResolutionScoreJob } from './resolution-score.worker';
+export { handleConflictDetectionJob } from './conflict-detection.worker';
+export { handleDailyDigestJob } from './daily-digest.worker';
+export { handleAIInsightNotificationsJob } from './insight-notifications.worker';
+export { handleStalenessCheckJob } from './staleness-check.worker';
+export { handleAIInlineJob } from './ai-inline.worker';
