+++ lib/queue/workers/staleness-check.worker.ts
@@ -1,0 +1,93 @@
+import { logger } from '@/lib/infrastructure/logger';
+import { prisma } from '@/lib/infrastructure/prisma';
+import type { StalenessCheckJobData } from '../types';
+
+const STALE_THRESHOLD_DAYS = 30;
+const RESOLUTION_SCORE_THRESHOLD = 50;
+const STALE_BATCH_SIZE = 100;
+
+function isStale(updatedAt: Date, resolutionScore: number | null): boolean {
+  const ageDays = (Date.now() - updatedAt.getTime()) / (1000 * 60 * 60 * 24);
+  if (ageDays < STALE_THRESHOLD_DAYS) return false;
+  return resolutionScore === null || resolutionScore < RESOLUTION_SCORE_THRESHOLD;
+}
+
+async function handleStalenessBatchCheck() {
+  let checked = 0;
+  let updated = 0;
+  let cursor: string | undefined;
+
+  while (true) {
+    const threads = await prisma.thread.findMany({
+      where: {
+        isOutdated: false,
+        updatedAt: { lt: new Date(Date.now() - STALE_THRESHOLD_DAYS * 24 * 60 * 60 * 1000) },
+        OR: [
+          { resolutionScore: null },
+          { resolutionScore: { lt: RESOLUTION_SCORE_THRESHOLD } },
+        ],
+        deletedAt: null,
+        ...(cursor ? { id: { gt: cursor } } : {}),
+      },
+      select: { id: true },
+      orderBy: { id: 'asc' },
+      take: STALE_BATCH_SIZE,
+    });
+
+    if (threads.length === 0) break;
+
+    await prisma.thread.updateMany({
+      where: { id: { in: threads.map((t) => t.id) } },
+      data: { isOutdated: true, lastVerifiedAt: new Date() },
+    });
+
+    updated += threads.length;
+    checked += threads.length;
+    cursor = threads[threads.length - 1].id;
+
+    if (threads.length < STALE_BATCH_SIZE) break;
+  }
+
+  logger.info(`[worker:ai] staleness batch check complete — ${checked} checked, ${updated} updated`);
+  return { handled: true, checked, updated };
+}
+
+export async function handleStalenessCheckJob(data: StalenessCheckJobData) {
+  const { threadId, cronJob } = data;
+
+  if (!cronJob && !threadId) {
+    throw new Error('Missing required fields: threadId or cronJob must be provided');
+  }
+
+  if (cronJob && !threadId) {
+    logger.info(`[worker:ai] staleness-check batch job`);
+    return handleStalenessBatchCheck();
+  }
+
+  logger.info(`[worker:ai] staleness-check for thread ${threadId}`);
+
+  const thread = await prisma.thread.findFirst({
+    where: { id: threadId!, deletedAt: null },
+    select: { id: true, updatedAt: true, resolutionScore: true, isOutdated: true },
+  });
+
+  if (!thread) {
+    logger.warn(`[worker:ai] Thread ${threadId} not found for staleness check`);
+    return { handled: true, checked: 1, updated: 0 };
+  }
+
+  if (thread.isOutdated) {
+    return { handled: true, checked: 1, updated: 0 };
+  }
+
+  if (isStale(thread.updatedAt, thread.resolutionScore)) {
+    await prisma.thread.update({
+      where: { id: threadId! },
+      data: { isOutdated: true, lastVerifiedAt: new Date() },
+    });
+    logger.info(`[worker:ai] Thread ${threadId} marked as outdated`);
+    return { handled: true, checked: 1, updated: 1 };
+  }
+
+  return { handled: true, checked: 1, updated: 0 };
+}
