+++ modules/messages/adapters/infra-side-effects.ts
@@ -1,5 +1,8 @@
 import { sendMentionNotification } from '@/lib/services/email';
 import { enqueueInlineJob } from '@/lib/services/queue';
+import { createBulkNotifications } from '@/modules/notifications/repository';
+import { recordActivity } from '@/modules/activity/repository';
+import { revalidatePath } from 'next/cache';
 import type { MessageSideEffectsPort } from '@/modules/messages/ports/side-effects';
 
 export const infraMessageSideEffects: MessageSideEffectsPort = {
@@ -8,5 +11,17 @@ export const infraMessageSideEffects: MessageSideEffectsPort = {
   },
   async enqueueAiInline(args) {
     await enqueueInlineJob(args);
+  },
+  async createBulkNotifications(notifications) {
+    await createBulkNotifications(notifications);
+  },
+  recordActivity(args) {
+    return recordActivity(args);
+  },
+  revalidateThreadPage(slug) {
+    if (slug) revalidatePath(`/dashboard/threads/${slug}`);
+  },
+  revalidateDashboard() {
+    revalidatePath('/dashboard');
   },
 };
