+++ modules/messages/ports/side-effects.ts
@@ -1,5 +1,6 @@
-// Injected into message actions so tests can assert on emails/jobs without
-// touching SMTP or QStash. See adapters/infra-side-effects.ts for the real one.
+// Injected into message actions so tests can assert on side effects without
+// touching SMTP, QStash, Redis, or Next.js cache. See adapters/infra-side-effects.ts
+// for the real one.
 export interface MessageSideEffectsPort {
   sendMentionEmail: (args: {
     toEmail: string;
@@ -14,4 +15,20 @@ export interface MessageSideEffectsPort {
     query: string;
     userId: string;
   }) => Promise<void>;
+  createBulkNotifications: (notifications: Array<{
+    userId: string;
+    type: 'MENTION' | 'SYSTEM';
+    title: string;
+    message: string;
+    data?: Record<string, unknown>;
+  }>) => Promise<void>;
+  recordActivity: (args: {
+    userId: string;
+    type: string;
+    entityType: string;
+    entityId: string;
+    metadata?: Record<string, unknown>;
+  }) => Promise<unknown>;
+  revalidateThreadPage: (slug: string | null) => void;
+  revalidateDashboard: () => void;
 }
