+++ modules/threads/ports/side-effects.ts
@@ -1,0 +1,10 @@
+import type { JobMessageData } from '@/lib/queue/types';
+
+type InitialMessage = Pick<JobMessageData, 'id' | 'content' | 'senderId' | 'sender' | 'createdAt'>;
+
+export interface ThreadSideEffectsPort {
+  enqueueInitialAiJobs: (args: {
+    threadId: string;
+    message: InitialMessage;
+  }) => Promise<void>;
+}
